refactor(concept-cat): Separate ESQueryUtil from ConceptSearchController
