+++ applications/concept-cat/src/main/java/no/ccat/controller/Common.java
@@ -1,0 +1,5 @@
+package no.ccat.controller;
+
+public class Common {
+    public static final String MISSING = "MISSING";
+}
