+++ applications/concept-cat/src/main/java/no/ccat/controller/ConceptSearchController.java
@@ -10,10 +10,7 @@ import no.fdk.webutils.aggregation.ResponseUtil;
 import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.index.query.RangeQueryBuilder;
-import org.elasticsearch.search.aggregations.AbstractAggregationBuilder;
 import org.elasticsearch.search.aggregations.AggregationBuilders;
-import org.elasticsearch.search.aggregations.bucket.filters.FiltersAggregator;
 import org.elasticsearch.search.aggregations.bucket.terms.Terms;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,11 +28,10 @@ import org.springframework.hateoas.PagedResources;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Arrays;
-import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 
-import static java.lang.Integer.MAX_VALUE;
+import static no.ccat.controller.Common.MISSING;
 import static org.apache.commons.lang3.StringUtils.isNotEmpty;
 
 @CrossOrigin
@@ -43,8 +39,6 @@ import static org.apache.commons.lang3.StringUtils.isNotEmpty;
 @RequestMapping(value = "/concepts")
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ConceptSearchController {
-    public static final String MISSING = "MISSING";
-    public static final long DAY_IN_MS = 1000 * 3600 * 24;
 
     private static final Logger logger = LoggerFactory.getLogger(ConceptSearchController.class);
     private ElasticsearchTemplate elasticsearchTemplate;
@@ -116,7 +110,7 @@ public class ConceptSearchController {
         BoolQueryBuilder composedQuery = QueryBuilders.boolQuery().must(searchQuery);
 
         if (!orgPath.isEmpty()) {
-            composedQuery.filter(QueryUtil.createTermQuery("publisher.orgPath", orgPath));
+            composedQuery.filter(ESQueryUtil.createTermQuery("publisher.orgPath", orgPath));
         }
 
         NativeSearchQuery finalQuery = new NativeSearchQueryBuilder()
@@ -173,11 +167,11 @@ public class ConceptSearchController {
                 .order(Terms.Order.count(false)));
         }
         if (selectedAggregationFields.contains("firstHarvested")) {
-            searchQuery.addAggregation(QueryUtil.createTemporalAggregation("firstHarvested", "harvest.firstHarvested"));
+            searchQuery.addAggregation(ESQueryUtil.createTemporalAggregation("firstHarvested", "harvest.firstHarvested"));
         }
 
         if (selectedAggregationFields.contains("publisher")) {
-            searchQuery.addAggregation(QueryUtil.createTermsAggregation("publisher", "publisher.id.keyword"));
+            searchQuery.addAggregation(ESQueryUtil.createTermsAggregation("publisher", "publisher.id.keyword"));
         }
         return searchQuery;
     }
@@ -186,38 +180,6 @@ public class ConceptSearchController {
         //In order for spring to not include Source or Remark when its parts are empty we need to null out the source object itself.
         for (ConceptDenormalized concept : concepts) {
             ConceptGetController.stripEmptyObject(concept);
-        }
-    }
-
-    static class QueryUtil {
-        static RangeQueryBuilder createRangeQueryFromXdaysToNow(int days, String dateField) {
-            long now = new Date().getTime();
-
-            return QueryBuilders.rangeQuery(dateField).from(now - days * DAY_IN_MS).to(now).format("epoch_millis");
-        }
-
-        static QueryBuilder createTermQuery(String term, String value) {
-            return value.equals(MISSING) ?
-                QueryBuilders.boolQuery().mustNot(QueryBuilders.existsQuery(term)) :
-                QueryBuilders.termQuery(term, value);
-        }
-
-
-        static AbstractAggregationBuilder createTemporalAggregation(String name, String dateField) {
-
-            return AggregationBuilders.filters(name,
-                new FiltersAggregator.KeyedFilter("last7days", QueryUtil.createRangeQueryFromXdaysToNow(7, dateField)),
-                new FiltersAggregator.KeyedFilter("last30days", QueryUtil.createRangeQueryFromXdaysToNow(30, dateField)),
-                new FiltersAggregator.KeyedFilter("last365days", QueryUtil.createRangeQueryFromXdaysToNow(365, dateField)));
-        }
-
-        static AbstractAggregationBuilder createTermsAggregation(String aggregationName, String field) {
-            return AggregationBuilders
-                .terms(aggregationName)
-                .missing(MISSING)
-                .field(field)
-                .size(MAX_VALUE) //be sure all theme counts are returned
-                .order(Terms.Order.count(false));
         }
     }
 }
